Improve the tokenizer + cleanup...
- Removed some leftover debug definitions from Program.cs. I forgot to
  remove them after I removed the debug library.
- Reworked SourceFile references to be the interface
  ITextReaderSourceProvider, because virtual file references are going to
  be important later on.
- Added support for language distinction to tokens and the tokeniser. I was
  originally planning to do this, but I cut my initial work before the
  first commit.

Now, it's time for some remarks!!

As much as I love C# as a language, I really hate a lot of the conventions
that Microsoft and the C# community have cooked up for it. XML comments are
by far the most egregious, but variable naming convention is secondplace.

I like using the m_ prefix on class members. It's clean, concise, and still
lets you immediate identify that a member of the class is being changed
rather than some local variable.

A single _, which is the convention Microsoft prefers, is quite ugly by
my standards. I simply do not like beginning variable names on
non-alphanumeric characters.

I do this for class fields without complaint, neither from myself nor the
linter in the case of non-public variables. But it feels wrong to do it for
properties (the little {get; set;} things), which are far more like methods
than data in their own right.

For public properties, I just bite the bullet and use Microsoft's
convention. It didn't make immediately make sense to me when I got into C#
programming, but it is actually fairly consistent; properties share the
same syntax as methods but with data operators instead of arguments, so the
same naming convention applies. Whenever I access one of these in a method,
I make sure to prefix `this.` to make it extra clear that it's a class
member which is being accessed.

For private properties, however, it would be nice to use my own conventions
if I had anything decent to use. The most I could think of is the
`m_` prefix, which, while nice, feels inappropriate for the method-like
functionality of C# properties. I think I still prefer it, since it's more
concise than the verbose `this.`, but it feels awkward.

But I'm just very nitpicky.

As a matter of fact, I fully intend on looking into writing a Visual Studio
extension so I can use JavaDoc-like comments in C# while maintaining the
convenient little documentation hovercards.

I'm not sure who at Microsoft thought it was a good idea to embed ugly
XML blobs into code comments meant to be read by humans, but I'm even more
perplexed by why they have stood by this design decision for so long, even
after reducing the verbosity of C# tenfold.

Did you know?? The most common string of characters in all C# files in the
world isn't any of the programming content words. It's not `using` or
`class` or `return` or anything of the sort. It is, in fact, the ugly-ass
XML comment blob "<summary>" which pollutes the C# language.

Countless gigabytes of space worldwide spent on storing these fucking
"<summary>" "</summary>" tags.

And yet Java, THE LANGUAGE WHICH INSPIRED C#, did it right back in 1996.
Sun realised that programming languages and comments exist for human
consumption first and foremost, and thus worked on designing a code-
documentation language which is reasonably human-readable at the expense
of being ever-so-slightly harder for your computer to make sense of.

I suppose Microsoft expects people to constantly run their code through
the documentation generator and pull up the pretty-printed documentation
to the side instead of just reading it in their code editor. That's the
only possible explanation for how this has gone unfixed for so long.

Whatever. When I figure out how to make this Visual Studio extension, I
will update the repo to replace all code comments with variants that are
pleasant to read inside the text editor.

Signed-off-by: Isabella <kawapure@gmail.com>