Push current working state...
Most of this is unfinished or will be restructured in the future.

This notably includes early work on parse nodes.

When I started this project, I actually forgot just how simple the C
preprocessor language actually is. It would be trivial to support almost
all of the possible operations, although I don't see any point in handling
something like #line because we don't need to support __FILE__ or __LINE__
in the first place here.

Preprocessor inclusion will be optimised if the file begins with
`#pragma once` or if the file starts with a condition.

I didn't initially plan to do this, but I plan to add support for including
additional DUIXML files inside of an existing one via #include. This is the
only case where file contents other than headers will not be ignored. There
is already some preliminary work done on this regarding file extensions.

Signed-off-by: Isabella <kawapure@gmail.com>