Support line/column offsets for debug messages...
This is a fun one.

As simple as it may seem, determining a meaningful offset for the user's
editor is actually a little complicated. This is, in particular, because of
the FUCKING tab character, which take up virtual space in the text editor.

The default tab size is 4 in every text editor I've used, except for the
built-in code editor of GitHub, which uses 8 by default for some reason.

Basically other than that, you can predict the text-editor representation
of a source code file pretty well.

Byte offsets are preferred by some users, so a parameter will be provided
to use those instead of line/column offsets via a startup parameter. See
this opinion article for more information:
https://www.computerenhance.com/p/byte-positions-are-better-than-line

Signed-off-by: Isabella <kawapure@gmail.com>